pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.33.0",
    "numpy>=2.3.4",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "streamlit>=1.51.0",
]
